+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- Introduce an install script for cogment CLI.
+
 ## v1.0.3 - 2021-07-07
 
 ### Changed